Avoid deprecation warnings
See https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

+++ algorithms/symmetry/cosym/test_cosym_target.py
@@ -20,7 +20,7 @@ def test_cosym_target(space_group):
         i += 1
         intensities = intensities.concatenate(d, assert_is_similar_symmetry=False)
         dataset_ids[i * d.size() : (i + 1) * d.size()] = np.full(
-            d.size(), i, dtype=np.int
+            d.size(), i, dtype=int
         )
 
     for weights in [None, "count", "standard_error"]:
